added controllers ,services

+++ src/main/java/ticketing/com/ticketingsystem/repositories/TicketPoolRepository.java
@@ -3,5 +3,5 @@ package ticketing.com.ticketingsystem.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ticketing.com.ticketingsystem.models.TicketPool;
 
-public interface TicketPoolRepository extends JpaRepository<TicketPool, Integer> {
+public interface TicketPoolRepository extends JpaRepository<TicketPool, Long> {
 }
