+++ src/main/java/ticketing/com/ticketingsystem/repositories/VendorRepository.java
@@ -3,5 +3,5 @@ package ticketing.com.ticketingsystem.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ticketing.com.ticketingsystem.models.Vendor;
 
-public interface VendorRepository extends JpaRepository<Vendor, Integer> {
+public interface VendorRepository extends JpaRepository<Vendor, Long> {
 }
